test: cover content bundle script
